telegram_bot/requirements.txt
python-telegram-bot
httpx
python-dotenv
pyzbar
Pillow
